+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto_classes/teleOpMecanum.java
@@ -4,12 +4,10 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.GyroSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
-
-import org.firstinspires.ftc.teamcode.team_methods.TeamMethods
-import org.firstinspires.ftc.teamcode.team_methods.Drive;
+import org.firstinspires.ftc.teamcode.team_methods.TeamMethods;
 
 @TeleOp(name="test", group="test") //fix this
-public class teleOpMecanum extends OpMode {
+public class teleOpMecanum extends TeamMethods {
     private ElapsedTime runtime = new ElapsedTime();
 
     //these values should be determined based off hardware being used
@@ -73,7 +71,7 @@ public class teleOpMecanum extends OpMode {
         double xNew = relativeValues[1];
         double yNew = relativeValues[2];
         double angleNew = relativeValues[3];
-        Drive.driveToPosition(xNew,yNew,angleNew, true);
+        driveToPosition(xNew,yNew,angleNew, true);
         telemetry.addData("Status", "Run Time: " + runtime.toString());
     }
