+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pathfinding/GameObject.java
@@ -1,0 +1,5 @@
+package org.firstinspires.ftc.teamcode.pathfinding;
+
+public class GameObject extends Obstacle{
+
+}
