+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pathfinding/Node.java
@@ -4,25 +4,18 @@ package org.firstinspires.ftc.teamcode.pathfinding;
 An class and object which is in essence a position marker.
  */
 public class Node {
-    double xpos;
-    double ypos;
+    double X_POS;
+    double Y_POS;
 
     //object
-    Node(double xposI, double yposI) {
-        xpos = xposI;
-        ypos = yposI;
+    Node(double X_POS_I, double Y_POS_I) {
+        X_POS = X_POS_I;
+        Y_POS = Y_POS_I;
     }
 
     //returns X and Y pos
     public double[] GetPos() {
-        double returnV[] = new double[]{xpos,ypos};
+        double returnV[] = new double[]{X_POS,Y_POS};
         return returnV;
     }
-
-    //instantiates several default nodes
-    public static void instantiateDefaultField() {
-
-    }
-
-
 }
